requirements.txt
# Requisitos principales del sistema (comentarios importantes)
# -------------------------------------------------------
# Para Windows: Necesitas Microsoft Visual C++ 14.0 o superior para scikit-surprise
# Descargar: https://visualstudio.microsoft.com/visual-cpp-build-tools/
# Python 3.8 o superior recomendado (funciona bien con la v.3.10.11)

# Paquetes base esenciales
# ------------------------
numpy==1.24.4  # Base numérica (requerido por pandas, scipy, etc.)
pandas==2.0.3   # Procesamiento de datos
scipy==1.15.3   # Operaciones científicas (requerido por surprise)

# Machine Learning y Recomendación
# -------------------------------
scikit-surprise==1.1.3  # Sistema de recomendación (requiere C++ en Windows)
joblib==1.4.2           # Para guardar/cargar modelos entrenados
scikit-learn==1.4.0     # Utilizado internamente por surprise

# Procesamiento de datos adicional
# --------------------------------
pyarrow==20.0.0         # Manejo eficiente de datos (usado por pandas)
python-dateutil==2.9.0  # Manejo de fechas
pytz==2025.2            # Zonas horarias
tzdata==2025.2          # Base de datos de zonas horarias

# Visualización
# -------------
matplotlib==3.8.4       # Gráficos básicos
seaborn==0.13.2         # Gráficos estadísticos
altair==5.5.0           # Visualizaciones interactivas

# Interfaz Web (Streamlit)
# ------------------------
streamlit==1.35.0        # Framework principal
pydeck==0.9.1           # Mapas para streamlit

# Utilidades y manejo de datos
# ----------------------------
Jinja2==3.1.6           # Plantillas (usado por Streamlit)
MarkupSafe==3.0.2        # Seguridad para Jinja2
protobuf==4.25.8         # Serialización (requerido por algunas dependencias)
tenacity==8.5.0          # Retry para operaciones

# Dependencias de desarrollo (opcionales)
# --------------------------------------
gitdb==4.0.12            # Para control de versiones con Git
GitPython==3.1.44        # Integración con Git
watchdog==6.0.0          # Recarga automática durante desarrollo